requirements/base.txt
Django==2.0.1
django-authtools==1.6.0
prettyconf==1.2.3
dj-database-url==0.4.2
psycopg2==2.7.3.2
pillow==5.0.0
django-anymail==1.2
django-extensions==1.9.9
whitenoise==3.3.1
django-widget-tweaks==1.4.1
djangorestframework==3.7.7
django-filter==1.1.0